plotMotorTickRateVsPower()
This visualization needs modifications.

TickRate is calculated using a simple diff() function,
which could be creating artifacts. It can limit the
maxRate output, but this isn't used.

Because the motors are in "USE_ENCODER" mode,
the Motor Power isn't really power, but the speed
 commanded to the motor PID controller.

Added analysis function plotMotorTickRateVsPower()
which needs work, and might not be not useful
at the moment. A heat map or some box plots might
be better.