Split module into parseCSV and telemetryAnalysis modules
parseCSV has a single method, getDictionary(filename)
which simply returns a dictionary of field arrays from 
the csv file. If no filename is provided, a file dialog 
is opened.

The telemetryAnalysis module provides a variety 
of analysis and visualization tools, and will run a 
set group of them if the module file is run directly.
Some of this code calls getDictionary() to capture 
the csv data.